tools: Allow for multiple release branch prefixes in a plugin (#29707)
We want to allow for releasing plugins in groups, for example the weekly
release should include both jetpack and mu-wpcom-plugin, and the monthly
should include jetpack, mu-wpcom-plugin, backup, and social.

However, as can be seen in the list above, we sometimes don't release
_all_ the plugins in a group. And when we don't, that makes some tests
fail because they can't tell the difference between "plugin in the group
didn't get updated for release" and "plugin is not being released".

The solution to this is to allow a plugin to be in multiple groups,
represented by it belonging to multiple prefixes.

This PR also removes the ability to release anything other than an
entire prefix via `create-release-branch.sh`: you can no longer list
multiple plugins on the command line, and you can no longer choose to
release just one plugin if the plugin is always grouped.

This also loosens the check for a release branch in `jetpack changelog`
and the pre-push hook, to accept release branches that use something
other than a version number after the `/branch-`.

`tools/get-contributor-list.sh` is rewritten to not rely on release
branches at all, instead it looks through the git history for the
commits where the top entry in CHANGELOG.md matches the target version.
There's a slight possibility for a race there to miss a contributor (if
their one contribution is merged in between the prerelease branch cut
and it getting merged back), but that seems fairly unlikely.

Note `tools/release-plugin.sh` is broken after this commit; we'll fix it
as part of #29323.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4632092861